feat: projects list page
