+++ astro-tmp/src/lib/ui/error-page.astro
@@ -1,15 +1,14 @@
 ---
 import Container from './container.astro'
-import { PostHeading } from './posts-layout/posts-heading'
+import { PostHeading } from './post-layout/post-heading'
 
 export interface Props {
   error: unknown
-  // reset: () => void
   tryAgain?: boolean
   statusCode?: number
 }
 
-const { error,  tryAgain = true, statusCode = 500 } = Astro.props
+const { error, tryAgain = true, statusCode = 500 } = Astro.props
 
 const DEFAULT_MESSAGE = 'I messed up :('
 ---
@@ -21,17 +20,20 @@ const DEFAULT_MESSAGE = 'I messed up :('
       {(error instanceof Error ? error.message : `${error}`) || DEFAULT_MESSAGE}
     </PostHeading>
 
-    <!-- TODO: add reset button -->
-    <!-- {
+    {
       tryAgain && (
         <>
-          <button class="group mx-auto my-8 block max-w-xs text-center text-lg">
+          <a
+            data-astro-reload
+            href={Astro.url}
+            class="group mx-auto my-8 block max-w-xs text-center text-lg"
+          >
             Click <span class="text-drac-pink group-hover:underline">here</span> to try again.
-          </button>
+          </a>
           <p class="text-center italic">or</p>
         </>
       )
-    } -->
+    }
     <a href="/" class="group mx-auto my-8 block max-w-xs text-center text-lg">
       Click <span class="text-drac-pink group-hover:underline">here</span> to go home.
     </a>
