+++ astro-tmp/src/lib/ui/post-layout/post-bottom-section.astro
@@ -1,0 +1,1 @@
+<div class="mx-auto mt-24! max-w-xs space-y-8 text-center text-sm"><slot /></div>
