+++ astro-tmp/src/lib/ui/post-layout/post-description.astro
@@ -1,0 +1,1 @@
+<p class="mt-6 mb-12 text-center text-lg text-pretty"><slot /></p>
